+++ resources/js/components/MostrarInfoPaciente.vue
